# Pre allocate Vec of correct size for community parsing