First, I added an example that just "scans" (iterates without touching any object) a RIB file using the record or element iterator. Then, when profiling, I noticed that a significant amount of time was spent on resizing specific Vecs, I pre-allocated Vecs of the corrrect size in community parsing.

±4-5% of CPU time was spent in re-sizing (and allocating) a Vec when parsing a RIB. This change pre-allocates a Vec of the correct size.

Note that there is a performance regression since 0.11.1 - which warrants a separate investigation.

Transcript of the performance change, using hyperfine on an otherwise completely idle  machine with AMD EPYC 7502P processor follows below.

```
Before:
$ hyperfine "./target/release/examples/scan_mrt ~/bview.20250101.0000.gz elements"
Benchmark 1: ./target/release/examples/scan_mrt ~/bview.20250101.0000.gz elements
  Time (mean ± σ):     70.117 s ±  1.019 s    [User: 69.762 s, System: 0.070 s]
  Range (min … max):   69.119 s … 72.314 s    10 runs

$ git stash pop
On branch feature/pre-allocate-vecs
Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
        modified:   src/parser/bgp/attributes/attr_08_communities.rs
        modified:   src/parser/bgp/attributes/attr_16_25_extended_communities.rs
        modified:   src/parser/bgp/attributes/attr_32_large_communities.rs

$ cargo build --release --examples --features=cli
   Compiling bgpkit-parser v0.11.1 (/home/tdekock/src/bgpkit-parser)
    Finished `release` profile [optimized] target(s) in 6.18s

After:
$ hyperfine "./target/release/examples/scan_mrt ~/bview.20250101.0000.gz elements"
Benchmark 1: ./target/release/examples/scan_mrt ~/bview.20250101.0000.gz elements
  Time (mean ± σ):     66.255 s ±  0.584 s    [User: 65.920 s, System: 0.068 s]
  Range (min … max):   65.716 s … 67.643 s    10 runs
```